[ntuple] Move string serialization functions to RNTupleUtil
Move the `{Serialize,Deserialize}String` functions to the
`ROOT::Experimental::Internal::RNTupleSerialization` namespace where they may be
reused by other parts of RNTuple.
